fix(diplomacy): split scenario fixtures to satisfy 400-line test cap (#3825)
Move GP-tribe and resolver scenario builders into a sibling support file
so repo.domain_package_test_file_size passes on PR #3830.
