+++ packages/colonizethis_diplomacy/test/support/diplomacy_game_fixtures.dart
@@ -1,2 +1,3 @@
 export 'diplomacy_game_fixtures_base.dart';
 export 'diplomacy_game_fixtures_scenarios.dart';
+export 'diplomacy_game_fixtures_scenarios_gp_tribe.dart';
